# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickFlow API",
    "description": "Tasks, habits, learning resources, todo plans, settings and dashboard for the QuickFlow personal productivity app.",
    "version": "v1"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Lists tasks. Archived tasks are excluded unless `archived` says otherwise.",
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "description": "Case-insensitive text contained in the title.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TaskItemStatus"
            }
          },
          {
            "name": "Priority",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TaskPriority"
            }
          },
          {
            "name": "DueDate",
            "in": "query",
            "description": "Only tasks due on this date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "DueFrom",
            "in": "query",
            "description": "Only tasks due on or after this date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "DueTo",
            "in": "query",
            "description": "Only tasks due on or before this date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Overdue",
            "in": "query",
            "description": "true = only overdue tasks, false = only tasks that are not overdue.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Archived",
            "in": "query",
            "description": "Exclude (default) hides archived tasks, Include shows all, Only shows archived tasks.",
            "schema": {
              "$ref": "#/components/schemas/ArchivedFilter"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "description": "CreatedAt (default) or DueDate. Tasks without a due date sort last.",
            "schema": {
              "$ref": "#/components/schemas/TaskSortBy"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "description": "Desc (default) or Asc.",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Creates a task.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Gets one task (archived tasks included).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Updates a task's title, description, status, priority and due date.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Permanently deletes a task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Marks a task as completed (status Done).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/archive": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Archives a task (hidden from the default list).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/restore": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Restores an archived task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArchivedFilter": {
        "enum": [
          "Exclude",
          "Include",
          "Only"
        ],
        "type": "string"
      },
...
      "CreateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Required, at most 200 characters.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Optional, at most 2000 characters.",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/TaskItemStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "dueDate": {
            "type": "string",
            "description": "Optional due date (yyyy-MM-dd).",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Body for creating a task."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ],
        "type": "string"
      },
      "TaskItemStatus": {
        "enum": [
          "Todo",
          "InProgress",
          "Done"
        ],
        "type": "string"
      },
      "TaskPriority": {
        "enum": [
          "Low",
          "Medium",
          "High"
        ],
        "type": "string"
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/TaskItemStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isArchived": {
            "type": "boolean"
          },
          "isOverdue": {
            "type": "boolean",
            "description": "True when the due date is before today and the task is not Done (and not archived)."
          }
        },
        "additionalProperties": false,
        "description": "A task as returned by the API."
      },
      "TaskSortBy": {
        "enum": [
          "CreatedAt",
          "DueDate"
        ],
        "type": "string"
      },
...
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/TaskItemStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Body for updating a task. Title is required. Omitted/null Status or Priority keep their\r\ncurrent value; Description and DueDate are replaced (null clears them)."
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      }
    }
  }
}